pyproject.toml
[tool.poetry]
name = "utility-wrapper"
version = "0.1.0"
description = "Wrapper service for useful utilities"
authors = ["QC Ware Engineering <engineering@qcware.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.135.1"}
uvicorn = "^0.41.0"
openbabel-wheel = "^3.1.1.23"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
